# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LOGISTIK UNIKOM API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/unit-types": {
            "get": {
                "tags": [
                    "Unit Type"
                ],
                "summary": "List unit type (support filter parent_id)",
                "description": "Jika parent_id tidak dikirim → tampilkan tree parent + children. Jika parent_id=null → hanya parent. Jika parent_id=id → children berdasarkan parent.",
                "operationId": "9576d5a63c1ad2d38100a0dee969cb30",
                "parameters": [
                    {
                        "name": "parent_id",
                        "in": "query",
                        "description": "Filter parent_id (isi null atau id)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "null"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List unit type berhasil diambil"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Unit Type"
                ],
                "summary": "Tambah unit type",
                "operationId": "c8f66de70c012772bd479cd5608615ef",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Fakultas"
                                    },
                                    "parent_id": {
                                        "type": "integer",
                                        "example": null,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Unit type berhasil dibuat"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/unit-types/{id}": {
            "get": {
                "tags": [
                    "Unit Type"
                ],
                "summary": "Detail unit type beserta children",
                "operationId": "c476703e1c68abb56815512d85cc99c2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID unit type",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail unit type berhasil diambil"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Unit Type"
                ],
                "summary": "Update unit type",
                "operationId": "612ca9893ee5d7d9854c25d50f3b156a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID unit type",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Program Studi"
                                    },
                                    "parent_id": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Unit type berhasil diupdate"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Unit Type"
                ],
                "summary": "Hapus unit type",
                "operationId": "1f64478aa05a0eca0ab989a4122a2568",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID unit type",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Unit type berhasil dihapus"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}